+++ SpatialGDK/Source/SpatialGDKEditor/Public/SpatialGDKEditorSchemaGenerator.h
@@ -12,6 +12,8 @@ SPATIALGDKEDITOR_API void ClearGeneratedSchema();
 
 SPATIALGDKEDITOR_API void DeleteGeneratedSchemaFiles();
 
+SPATIALGDKEDITOR_API void CopyWellKnownSchemaFiles();
+
 SPATIALGDKEDITOR_API bool TryLoadExistingSchemaDatabase();
 
 SPATIALGDKEDITOR_API bool GeneratedSchemaFolderExists();