+++ SpatialGDK/Source/SpatialGDKServices/Private/SpatialGDKServicesModule.cpp
@@ -1,0 +1,19 @@
+// Copyright (c) Improbable Worlds Ltd, All Rights Reserved
+
+#include "SpatialGDKServicesModule.h"
+#include "SpatialGDKServicesPrivate.h"
+
+#define LOCTEXT_NAMESPACE "FSpatialGDKServicesModule"
+
+DEFINE_LOG_CATEGORY(LogSpatialGDKServices);
+
+IMPLEMENT_MODULE(FSpatialGDKServicesModule, SpatialGDKServices);
+
+void FSpatialGDKServicesModule::StartupModule()
+{
+}
+void FSpatialGDKServicesModule::ShutdownModule()
+{
+}
+
+#undef LOCTEXT_NAMESPACE